backend/requirements.txt
Django==6.0.4
asgiref==3.11.1
dj-database-url==3.1.2
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
psycopg==3.3.3
psycopg-binary==3.3.3
PyJWT==2.12.1
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2026.2
